removed bar_chart code from app.py
