+++ src/app.py
@@ -1,8 +1,9 @@
 from dash import Dash, html, dcc, Input, Output, callback,callback_context
 from dash.dependencies import ALL
 import dash_bootstrap_components as dbc
-import altair as alt
-import dash_vega_components as dvc
+from components import bar_chart
+import callbacks
+
 import dash_leaflet as dl
 import pandas as pd
 import ast
@@ -120,42 +121,6 @@ park_map = dl.Map(
     zoomControl=True
 )
 
-def create_bar_chart(data):
-    parks_chart_filter = data.groupby("FacilityType").sum().reset_index()[["FacilityType", "FacilityCount"]]
-    parks_chart_filter = parks_chart_filter.sort_values(by="FacilityCount", ascending=False)[:5]
-    alt.theme.enable("fivethirtyeight")
-
-    chart = alt.Chart(parks_chart_filter).mark_bar().encode(
-        x=alt.X('FacilityCount:Q').scale(nice=True).axis(format='d'),
-        y=alt.Y('FacilityType:N').sort('-x'),
-        color=alt.Color('FacilityType', legend=None),
-        tooltip = ['FacilityType','FacilityCount']
-    ).configure(
-        background="transparent",
-    ).configure_axis(
-        labelColor= "black",
-        titleColor= "grey",
-        labelFontSize=12,
-        titleFontSize=12,
-        labelFontStyle='Helvetica'
-    ).properties(
-        title="Top 5 Facilities",
-        width=550,
-        height=80
-    ).configure_title(
-        font='Helvetica',
-        fontSize=16,
-        anchor='start',
-    )
-
-    return chart.to_dict()
-
-bar_chart = dvc.Vega(
-    id="bar-chart",
-    spec=create_bar_chart(facilities_data),
-    style={'width': '100%'}
-)
-
 # Define the layout with a map centered on Vancouver
 app.layout = dbc.Container(fluid=True, children=[
     dbc.Row([
@@ -245,49 +210,6 @@ def update_map(selected_neighbourhood, selected_facilities, selected_special_fea
     avg_hectare_text = f"{avg_hectare_filtered:.2f}" if not df_filtered.empty else "0.00"
 
     return [dl.TileLayer()] + create_markers(df_filtered) if not df_filtered.empty else [dl.TileLayer()], num_parks_text, avg_hectare_text
-
-@app.callback(
-    Output("bar-chart", "spec"),
-    Input("neighbourhood-dropdown", "value"),
-    Input("facility-dropdown", "value"),
-    Input("special-feature-dropdown", "value"),
-    Input("washrooms-checkbox", "value")
-)
-def update_bar_chart(selected_neighbourhood, selected_facilities, selected_special_features, washroom_filter):
-    df_filtered = parks_data.copy()
-
-    # Filter by neighbourhood
-    if selected_neighbourhood:
-        df_filtered = df_filtered[df_filtered["NeighbourhoodName"] == selected_neighbourhood]
-
-    # Filter by washroom availability
-    if washroom_filter is None:
-        washroom_filter = []
-    
-    if "Y" in washroom_filter:
-        df_filtered = df_filtered[df_filtered["Washrooms"] == "Y"]
-
-    park_ids = set(df_filtered["ParkID"])
-
-    # Ensure parks contain *all* selected facilities
-    if selected_facilities:
-        facility_counts = facilities_data[facilities_data["FacilityType"].isin(selected_facilities)]
-        facility_counts = facility_counts.groupby("ParkID")["FacilityType"].nunique()
-        matching_facilities = set(facility_counts[facility_counts == len(selected_facilities)].index)
-        park_ids &= matching_facilities if park_ids else matching_facilities  
-
-    # Ensure parks contain *all* selected special features
-    if selected_special_features:
-        special_counts = special_data[special_data["SpecialFeature"].isin(selected_special_features)]
-        special_counts = special_counts.groupby("ParkID")["SpecialFeature"].nunique()
-        matching_specials = set(special_counts[special_counts == len(selected_special_features)].index)
-        park_ids &= matching_specials if park_ids else matching_specials  
-
-    # Only filter for specific ParkIDs
-    facilities_data_new = facilities_data[facilities_data["ParkID"].isin(park_ids)]
-
-    # Update Chart
-    return create_bar_chart(facilities_data_new)
 
 @app.callback(
     Output("park-info", "children"),
